Optimize
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@3188 111d7451-93fa-4e81-90d4-8039f76616e9

+++ opcodes.h
@@ -47,7 +47,7 @@ typedef enum Reg_types {
 struct cpu_s {
     const char *name;
     const uint32_t *mnemonic;
-    const uint8_t *opcode;
+    const uint16_t *opcode;
     const uint16_t *disasm;
     const uint8_t *alias;
     const uint32_t registers;
